v0.2
Add hypotheses and cross-assignment between functions and hypotheses;
(temporary?) Flatten the intermediate function representation so that all of the arguments within the single-line stack are available within the single entity;
Add "breakpoints" for variables for the assignment calls that represent changes in the dataset;
Rework visualization into the tree graph with branching and color-coding for distinct hypotheses being explored. Color-coding is independent, branching occurs after each "breakpoint" on the dataset.